[Feat] RecordDetailUseCase 생성
