+++ MateRunner/MateRunner/Domain/UseCase/Record/Protocol/RecordDetailUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  RecordDetailUseCase.swift
+//  MateRunner
+//
+//  Created by 이정원 on 2021/11/23.
+//
+
+import Foundation
+
+protocol RecordDetailUseCase {
+    var runningResult: RunningResult { get }
+}
